requirements.txt
Django
django-mptt
django-model-utils
ipython